modified:   src/menus/scores.rs 	modified:   src/menus/start.rs 	modified:   src/menus/start_layout.rs 	modified:   src/utilities/score_board.rs 	modified:   src/utilities/score_board/file.rs 	renamed:    src/utilities/score_board/default_layout.rs -> src/utilities/score_board/score_layout.rs 	modified:   src/utilities/settings.rs 	modified:   src/utilities/settings/edit.rs 	modified:   src/utilities/settings/file.rs 	renamed:    src/utilities/settings/default_layout.rs -> src/utilities/settings/settings_layout.rs
