+++ src/utilities/settings.rs
@@ -1,4 +1,4 @@
-pub mod default_layout;
+pub mod settings_layout;
 pub mod defaults;
 pub mod edit;
 pub mod file;
